requirements.txt
fastapi[all]
psycopg2-binary
pytest